mr_subsampling/unet/requirements.txt
#atomicwrites==1.2.1
#attrs==18.2.0
certifi==2019.11.28
cffi==1.13.2
cloudpickle==1.3.0
cycler==0.10.0
dask==2.12.0
decorator==4.4.2
#h5py==2.10.0
imageio==2.8.0
kiwisolver==1.2.0
matplotlib==3.2.1
#more-itertools==4.3.0
networkx==2.4
numpy==1.18.2
olefile==0.46
Pillow==6.2.1
pluggy==0.8.0
protobuf==3.6.1
py==1.7.0
pycparser==2.19
pyparsing==2.4.6
#pytest==4.0.0
python-dateutil==2.8.1
pytz==2019.3
#PyWavelets==1.1.1
runstats==1.8.0
scikit-image==0.16.2
scipy==1.3.0
six==1.13.0
tensorboard==1.14.0
toolz==0.10.0
tornado==6.0.4
torch==1.3.1
torchvision==0.4.2
pytorch-lightning==0.7.3
test-tube==0.7.3
